Implement module loading
Ultimately I would like to have standard modules available, but as a
step on the way, this implements module loading that assumes files are
relative to the current working directory, and demonstrates its use
with the incumbent Kubernetes manifest generation.
