+++ nginx.js
@@ -1,6 +1,6 @@
-import k from 'kubernetes';
-import jk from 'jk';
+import k from 'kubernetes.js';
+import write from 'write.js';
 
 const container = k.Container('nginx', 'nginx:1.15.4');
 const deployment = k.Deployment('nginx', 3, [container]);
-jk.write(deployment);
+write(deployment);
